endpoint: Refactor host endpoint label initialization
Refactor the initialization of the host endpoint's labels as this code
will be used in another method in a subsequent commit.
